# docs/api/openapi.yaml
...
openapi: 3.1.0
info:
  title: KWG Directory API
  version: 0.1.0
  description: OpenChain Korea Work Group 멤버 주소록 REST API. 인증은 next-auth 세션 쿠키 기반. /api/auth/* 는 next-auth 내부 핸들러로 본 문서 범위 밖.
  license:
    name: Apache-2.0
servers:
  - url: /
    description: 동일 오리진
...
paths:
...
  /api/admin/members:
    get:
      summary: '[admin] 전체 멤버(대기 포함) 목록.'
      tags:
        - admin
      security:
        - sessionCookie: []
      responses:
        '200':
          description: 멤버 목록
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberList'
        '401':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      summary: '[admin] 멤버 승인/거절 토글.'
      tags:
        - admin
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminMembers'
      responses:
        '200':
          description: 처리됨
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminActionResult'
        '400':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/admin/approve:
    post:
      summary: '[admin] 가입 승인 + 메일링 초대.'
      tags:
        - admin
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminId'
      responses:
        '200':
          description: 승인됨
        '400':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/admin/admins:
    get:
      summary: '[admin] 어드민 목록.'
      tags:
        - admin
      security:
        - sessionCookie: []
      responses:
        '200':
          description: 어드민 목록
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberList'
        '401':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      summary: '[admin] 어드민 추가/제거.'
      tags:
        - admin
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminAdmins'
      responses:
        '200':
          description: 처리됨
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminActionResult'
        '400':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/admin/activity:
    get:
      summary: '[admin] 감사 활동 로그.'
      tags:
        - admin
      security:
        - sessionCookie: []
      responses:
        '200':
          description: 활동 로그
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityLog'
        '401':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/admin/export/csv:
    get:
      summary: '[admin] 멤버 CSV 내보내기.'
      tags:
        - admin
      security:
        - sessionCookie: []
      responses:
        '200':
          description: CSV
          content:
            text/csv:
              schema:
                type: string
        '401':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: authjs.session-token
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Member:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
        name_ko:
          type: string
        name_en:
          type: string
          nullable: true
        company:
          type: string
        role:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        category:
          type: string
          enum:
            - 기업
            - 연구/공공
            - 학계
          nullable: true
        email:
          type: string
          nullable: true
          description: 비로그인 시 redact
        email_public:
          type: boolean
        linkedin:
          type: string
          nullable: true
        github:
          type: string
          nullable: true
        discord:
          type: string
          nullable: true
        blog:
          type: string
          nullable: true
        avatar_url:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        approved:
          type: boolean
        created_at:
          type: string
          format: date-time
      required:
        - id
        - name_ko
        - company
        - approved
    MemberList:
      type: array
      items:
        $ref: '#/components/schemas/Member'
...
    AdminId:
      type: object
      description: zod AdminIdSchema 와 동기화.
      properties:
        id:
          type: string
          format: uuid
      required:
        - id
...
    AdminMembers:
      type: object
      description: zod AdminMembersSchema 와 동기화.
      properties:
        id:
          type: string
          format: uuid
        approved:
          type: boolean
      required:
        - id
        - approved
    AdminAdmins:
      type: object
      description: zod AdminAdminsSchema 와 동기화.
      properties:
        action:
          type: string
          enum:
            - add
            - remove
        user_id:
          type: string
          format: uuid
      required:
        - action
        - user_id
    AdminActionResult:
      type: object
      properties:
        success:
          type: boolean
        action:
          type: string
      required:
        - success
        - action
    ActivityLog:
      type: array
      items:
        type: object
...